requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyserial==3.5
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
python-dateutil==2.9.0.post0
shiboken6==6.8.1
six==1.17.0
